backend-python/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
openai==1.14.0
httpx==0.25.1
python-multipart==0.0.6
openai-agents==0.3.0